requirements.txt
absl-py==0.11.0
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.2.3
boto3==1.16.62
botocore==1.19.62
Brotli==1.0.9
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cycler==0.10.0
dash==1.19.0
dash-bootstrap-components==0.11.1
dash-canvas==0.1.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
debtcollector==2.2.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.2
Flask-Caching==1.9.0
Flask-Compress==1.8.0
flatbuffers==1.12
future==0.18.2
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
image-utils==0.1.6
imageai==2.1.6
imagecodecs==2021.1.28
imageio==2.9.0
importlib-metadata==3.4.0
ipykernel==5.4.3
ipynb-py-convert==0.4.6
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
iso8601==0.1.13
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-http-over-ws==0.0.8
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
keras-resnet==0.2.0
kiwisolver==1.3.1
lesscpy==0.14.0
llvmlite==0.35.0
lxml==4.6.2
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.2
mistune==0.8.4
mpmath==1.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
netaddr==0.8.0
netifaces==0.10.9
networkx==2.5
nibabel==3.2.1
nilearn==0.7.0
nipy==0.4.2
notebook==6.2.0
numba==0.52.0
numpy==1.19.3
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
oslo.i18n==5.0.1
oslo.utils==4.7.0
packaging==20.8
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
patsy==0.5.1
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
plotly==4.14.3
plotly-express==0.4.1
ply==3.11
progressbar==2.5
prometheus-client==0.9.0
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.1.3
pycparser==2.20
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-cephlibs==0.94.5.post1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.0
qtconsole==5.0.2
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.7
s3transfer==0.3.4
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.4.1
Send2Trash==1.5.0
shapes==0.1.3
silence-tensorflow==1.1.1
six==1.15.0
statsmodels==0.12.1
sympy==1.7.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2021.1.14
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.0
